requirements.txt
pymysql
mysql-connector-python
python-dotenv
pytest
pandas
matplotlib